refactor: unify the duplicated "Environment" class
RHCLOUD-23484
