+++ backend/src/test/java/com/redhat/cloud/notifications/routers/internal/InternalResourceTest.java
@@ -8,9 +8,9 @@ import com.redhat.cloud.notifications.db.ResourceHelpers;
 import com.redhat.cloud.notifications.models.Application;
 import com.redhat.cloud.notifications.models.BehaviorGroup;
 import com.redhat.cloud.notifications.models.Bundle;
+import com.redhat.cloud.notifications.models.Environment;
 import com.redhat.cloud.notifications.routers.dailydigest.TriggerDailyDigestRequestDto;
 import com.redhat.cloud.notifications.routers.engine.DailyDigestService;
-import com.redhat.cloud.notifications.routers.environment.Environment;
 import io.quarkus.test.common.QuarkusTestResource;
 import io.quarkus.test.junit.QuarkusTest;
 import io.quarkus.test.junit.mockito.InjectMock;
