+++ engine/src/main/java/com/redhat/cloud/notifications/templates/TemplateService.java
@@ -4,10 +4,10 @@ import com.redhat.cloud.notifications.config.FeatureFlipper;
 import com.redhat.cloud.notifications.ingress.Action;
 import com.redhat.cloud.notifications.ingress.Context;
 import com.redhat.cloud.notifications.ingress.Payload;
+import com.redhat.cloud.notifications.models.Environment;
 import com.redhat.cloud.notifications.recipients.User;
 import com.redhat.cloud.notifications.templates.extensions.ActionExtension;
 import com.redhat.cloud.notifications.templates.extensions.LocalDateTimeExtension;
-import com.redhat.cloud.notifications.templates.models.Environment;
 import io.quarkus.qute.Engine;
 import io.quarkus.qute.EvalContext;
 import io.quarkus.qute.ReflectionValueResolver;
