+++ engine/src/test/java/com/redhat/cloud/notifications/processors/email/aggregators/RhosakEmailAggregatorTest.java
@@ -6,8 +6,8 @@ import com.redhat.cloud.notifications.ingress.Event;
 import com.redhat.cloud.notifications.ingress.Metadata;
 import com.redhat.cloud.notifications.ingress.Payload;
 import com.redhat.cloud.notifications.models.EmailAggregation;
+import com.redhat.cloud.notifications.models.Environment;
 import com.redhat.cloud.notifications.templates.Rhosak.Templates;
-import com.redhat.cloud.notifications.templates.models.Environment;
 import io.quarkus.qute.TemplateInstance;
 import io.quarkus.test.junit.QuarkusTest;
 import io.vertx.core.json.JsonObject;
