+++ engine/src/test/java/com/redhat/cloud/notifications/templates/IntegrationsTemplatesTest.java
@@ -4,7 +4,7 @@ import com.redhat.cloud.notifications.TestHelpers;
 import com.redhat.cloud.notifications.config.FeatureFlipper;
 import com.redhat.cloud.notifications.ingress.Action;
 import com.redhat.cloud.notifications.models.Endpoint;
-import com.redhat.cloud.notifications.templates.models.Environment;
+import com.redhat.cloud.notifications.models.Environment;
 import io.quarkus.qute.TemplateInstance;
 import io.quarkus.test.junit.QuarkusTest;
 import org.junit.jupiter.api.BeforeEach;
