+++ engine/src/test/java/com/redhat/cloud/notifications/templates/models/EnvironmentTest.java
@@ -1,5 +1,6 @@
 package com.redhat.cloud.notifications.templates.models;
 
+import com.redhat.cloud.notifications.models.Environment;
 import io.quarkus.test.junit.QuarkusTest;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
